requirements.txt
appium-python-client>=3.0.0
selenium>=4.8.0
pytest>=7.0.0
pytest-html>=3.2.0
pytest-xdist>=3.0.0
allure-pytest>=2.9.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=9.0.0
requests>=2.28.0
pyyaml>=6.0
colorlog>=6.0.0